use jasmine for unit tests
