+++ package.json
@@ -16,10 +16,10 @@
     "build:zip": "zip-build themes/kleiderordnung dist --name kleiderordnung --template kleiderordnung.zip",
     "build:translations": "msgfmt -o src/languages/en_US.mo src/languages/en_US.po && msgfmt -o src/languages/de_DE.mo src/languages/de_DE.po",
     "watch": "watch \"npm run build:postscss\" src/css",
-    "test": "npm run stylelint && npm run eslint && npm run codeceptjs",
+    "test": "npm run stylelint && npm run eslint && npm run jasmine && npm run codeceptjs",
     "stylelint": "stylelint src/style.css src/css/*.css",
     "eslint": "eslint --ext .js src/js src/js/inc src/js/tests",
-    "jest": "jest",
+    "jasmine": "jasmine",
     "codeceptjs": "npm run codeceptjs:default && npm run codeceptjs:mobile",
     "codeceptjs:default": "codeceptjs run --steps",
     "codeceptjs:mobile": "MOBILE=true codeceptjs run --steps",
@@ -29,6 +29,7 @@
     "type": "git",
     "url": "git+https://github.com/openmindculture/kleiderordnung-berlin.git"
   },
+  "type": "module",
   "author": "Ingo Steinke",
   "license": "proprietary",
   "bugs": {
@@ -46,10 +47,10 @@
     "copyfiles": "^2.4.1",
     "docker-compose": "^0.24.0",
     "esbuild": "^0.19.2",
-    "eslint": "^8.36.0",
+    "eslint": "^8.48.0",
     "eslint-plugin-es5": "^1.5.0",
     "html-minifier-cli": "^0.1.0",
-    "jest": "^29.6.4",
+    "jasmine": "^5.1.0",
     "make-dir-cli": "^3.0.0",
     "npm-run-all": "^4.1.5",
     "path": "^0.12.7",
